# perf(lanelet2_extension): improve performance of getClosestLanelet