Signed-off-by: Taiga Takano [taiga.takano@tier4.jp](mailto:taiga.takano@tier4.jp)

## Description

If the size of the candidate_lanelets is 1, the "Find by Angle" task can be skipped.

## Before and After
This change speeds up the process as shown in the table below.

getClosestLanelet callback latency
|  PERCENTILE  |  BEFORE (us)  |  AFTER (us)  |
| ---- | ---- | ---- |
|  99%  |  42.5  |  33.3  |
|  90%  |  28.6  |  21.4  |
|  50%  |  15.7  |  10.3  |